requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.7
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-push-notifications==3.2.1
django-timezone-field==7.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.1
msgpack==1.1.0
packaging==24.2
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
six==1.17.0
sqlparse==0.5.3
stripe==11.6.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
